pyproject.toml
[project]
name = "RAGspert"
version = "0.7.0"
description = "Your async FastAPI starter Kit; without the fluff!"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiosqlite>=0.21.0",
    "asyncpg>=0.30.0",
    "crawl4ai>=0.7.4",
    "fastapi>=0.118.2",
    "google-genai>=1.43.0",
    "pgvector>=0.4.1",
    "pydantic-ai-slim[google]>=1.0.17",
    "pydantic-settings>=2.11.0",
    "sqlalchemy>=2.0.43",
    "uvicorn[standard]>=0.37.0",
]
